requirements.txt
altair==4.2.2
annotated-types==0.7.0
attrs==24.2.0
backoff==2.2.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.3
entrypoints==0.4
great-expectations==1.2.1
idna==3.10
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==3.0.2
marshmallow==3.23.1
mistune==3.0.2
monotonic==1.6
numpy==1.26.4
packaging==24.1
paho-mqtt==2.1.0
pandas==2.1.4
posthog==2.5.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.1
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.12
scipy==1.14.1
six==1.16.0
toolz==1.0.0
tqdm==4.66.6
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.2.3
